pipeline tasks now run in parallel
