+++ nlp/api.py
@@ -6,7 +6,8 @@ from claritynlp_logging import log, setup_log, ERROR, DEBUG
 
 import sys
 import signal
-from luigi_tools import luigi_runner
+#from luigi_tools import luigi_runner
+#from luigi_module import shutdown_workers
 
 def create_app(config_filename=None):
 
@@ -34,7 +35,8 @@ CORS(application)
 # handler for CTRL-C, to shutdown worker threads
 def _ctrl_c_handler(signal, frame):
     log('CTRL-C PRESSED')
-    luigi_runner.shutdown_workers()
+    #luigi_runner.shutdown_workers()
+    #shutdown_workers()
     sys.exit(0)
 
 # install the singal handler for CTRL-C
